cud 완성

+++ back/src/main/java/com/rebu/feed/review/validation/validator/ReviewRatingValidator.java
@@ -5,10 +5,10 @@ import com.rebu.feed.review.validation.annotation.ReviewRating;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class ReviewRatingValidator implements ConstraintValidator<ReviewRating, Long> {
+public class ReviewRatingValidator implements ConstraintValidator<ReviewRating, Integer> {
 
     @Override
-    public boolean isValid(Long rating, ConstraintValidatorContext constraintValidatorContext) {
+    public boolean isValid(Integer rating, ConstraintValidatorContext constraintValidatorContext) {
         if(rating == null || rating < 1 || rating > 5)
             throw new ReviewRatingMismatchException();
         return true;
